Index: frysk-core/frysk/proc/ChangeLog 2005-10-30  Andrew Cagney  <cagney@redhat.com>
	* TestProcGet.java: Fix typo, auxv.val not auxv.type.
	* Proc.java: Add .sendrecAuxv.  Implement .getAuxv locally.
	* LinuxProc.java: Replace .getAuxv with .sendrecAuxv.  Use
	frysk.sys.proc.AuxiliaryVectorBuilder to construct the auxv.
	* Auxv.java: Simplify.

+++ frysk-sys/frysk/sys/proc/cni/slurp.hxx
@@ -37,4 +37,8 @@
 // version and license this file solely under the GPL without
 // exception.
 
-extern long slurp (int pid, const char *name, char buf[], long sizeof_buf);
+/**
+ * Slurp <tt>/proc/PID/NAME</tt> into BUF (max size sizeof_buf).
+ * Return the actual number of byte read.
+ */
+extern int slurp (int pid, const char *name, char buf[], long sizeof_buf);
